LPS-187293 portal-search-admin-web: move methods to IndexActionsDisplayContextBuilder
https://liferay.atlassian.net/browse/LPS-187293
